+++ technical/sheet1/Factorial.java
@@ -7,11 +7,11 @@ public class Factorial {
     {
         System.out.print("Enter a number to find factorial:");
         Scanner scans = new Scanner(System.in);
-        byte num = scans.nextByte();
+        long num = scans.nextLong();
         System.out.println("factorial is: " + factorial(num));
         scans.close();
     }
-    static byte factorial(byte num) {
+    static byte factorial(long num) {
         byte i = 1;
        while(num != 1){
            i*=num;
